+++ packages/www/src/MDX/Pre/allComponents.ts
@@ -71,6 +71,7 @@ import {
   Flex,
   FlexItem,
   Form,
+  Grid,
   Heading,
   Icon,
   IconButton,
@@ -110,6 +111,8 @@ import {
   Select,
   SelectMulti,
   Slider,
+  Space,
+  SpaceVertical,
   Spinner,
   Swatch,
   Tab,
@@ -141,12 +144,16 @@ import styled from 'styled-components'
 import isEqual from 'lodash/isEqual'
 import { useState } from 'react'
 
+import { GridPlaceholder } from '../../helpers/GridPlaceholder'
+
 export const allComponents = {
   useState,
 
   styled,
 
   isEqual,
+
+  GridPlaceholder,
 
   palette,
   ActionList,
@@ -195,6 +202,7 @@ export const allComponents = {
   Flex,
   FlexItem,
   Form,
+  Grid,
   Heading,
   Icon,
   IconButton,
@@ -234,6 +242,8 @@ export const allComponents = {
   Select,
   SelectMulti,
   Slider,
+  Space,
+  SpaceVertical,
   Spinner,
   Swatch,
   Tab,