+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Layout components
+  - `Grid` component
+  - `Space` & `SpaceVertical` components
 - `SelectMulti` component
 - `CodeBlock` component
 - `InputTimeSelect` component
@@ -20,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ActionListManager` component
 - `Pagination` component
 - `ComponentsProvider` component to provide a pre-packaged `ThemeProvider` + `theme`
+- New Icons `VisibilityOutline`, `SendEmail`, `SendSftp`, `SendWebhook`
 
 ### Changed